core/local: Extract ./chokidar_steps/prepare_events
Including ChokidarWatcher#prepareEvents() & #oldMetadata()
